Wrap the kubelet drain in a conditional
All commands we're calling in the kubelet drain script are dependant on
the presence of a master node. In the case of a deployment deletion, it
might be that this node won't be present, which is why they're now
wrapped in a conditional that checks if a master node is still there.
For this, I reused a method that was still present in the commit 79b6618